--- amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/apps/recommendations/datasource/ProfileAppRecommendationsSubAssembler.kt
@@ -1,45 +1,0 @@
-/*
- * Copyright (c) 2025 Vitor Pamplona
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to use,
- * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
- * Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
- * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
- * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.apps.recommendations.datasource
-
-import com.vitorpamplona.amethyst.commons.relayClient.eoseManagers.PerUserEoseManager
-import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
-import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
-import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-
-class ProfileAppRecommendationsSubAssembler(
-    client: INostrClient,
-    allKeys: () -> Set<ProfileAppRecommendationsQueryState>,
-) : PerUserEoseManager<ProfileAppRecommendationsQueryState>(client, allKeys) {
-    override fun user(key: ProfileAppRecommendationsQueryState) = key.account.userProfile()
-
-    override fun updateFilter(
-        key: ProfileAppRecommendationsQueryState,
-        since: SincePerRelayMap?,
-    ): List<RelayBasedFilter> =
-        filterMyAppRecommendations(
-            pubkey = user(key).pubkeyHex,
-            relays = key.account.outboxRelays.flow.value,
-        ) +
-            filterRecentAppDefinitions(
-                relays = key.account.defaultGlobalRelays.flow.value,
-            )
-}
