--- amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/home/datasource/nip65Follows/FilterHomePostsByAuthors.kt
@@ -1,212 +1,0 @@
-/*
- * Copyright (c) 2025 Vitor Pamplona
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to use,
- * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
- * Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
- * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
- * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.home.datasource.nip65Follows
-
-import com.vitorpamplona.amethyst.commons.model.topNavFeeds.noteBased.author.AuthorsTopNavPerRelayFilterSet
-import com.vitorpamplona.amethyst.commons.model.topNavFeeds.noteBased.muted.MutedAuthorsTopNavPerRelayFilterSet
-import com.vitorpamplona.amethyst.commons.relayClient.subscriptions.ExplainedFilter
-import com.vitorpamplona.amethyst.commons.relayClient.subscriptions.SubPurpose
-import com.vitorpamplona.amethyst.commons.relays.SincePerRelayMap
-import com.vitorpamplona.quartz.experimental.attestations.attestation.AttestationEvent
-import com.vitorpamplona.quartz.experimental.birdstar.BirdDetectionEvent
-import com.vitorpamplona.quartz.experimental.birdstar.BirdexEvent
-import com.vitorpamplona.quartz.experimental.ephemChat.chat.EphemeralChatEvent
-import com.vitorpamplona.quartz.experimental.interactiveStories.InteractiveStoryPrologueEvent
-import com.vitorpamplona.quartz.experimental.nipsOnNostr.NipTextEvent
-import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
-import com.vitorpamplona.quartz.nip01Core.core.HexKey
-import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
-import com.vitorpamplona.quartz.nip18Reposts.GenericRepostEvent
-import com.vitorpamplona.quartz.nip18Reposts.RepostEvent
-import com.vitorpamplona.quartz.nip22Comments.CommentEvent
-import com.vitorpamplona.quartz.nip23LongContent.LongTextNoteEvent
-import com.vitorpamplona.quartz.nip35Torrents.TorrentEvent
-import com.vitorpamplona.quartz.nip53LiveActivities.chat.LiveActivitiesChatMessageEvent
-import com.vitorpamplona.quartz.nip53LiveActivities.streaming.LiveActivitiesEvent
-import com.vitorpamplona.quartz.nip54Wiki.WikiNoteEvent
-import com.vitorpamplona.quartz.nip64Chess.challenge.offer.LiveChessGameChallengeEvent
-import com.vitorpamplona.quartz.nip64Chess.end.LiveChessGameEndEvent
-import com.vitorpamplona.quartz.nip64Chess.game.ChessGameEvent
-import com.vitorpamplona.quartz.nip68Picture.PictureEvent
-import com.vitorpamplona.quartz.nip71Video.VideoHorizontalEvent
-import com.vitorpamplona.quartz.nip71Video.VideoNormalEvent
-import com.vitorpamplona.quartz.nip71Video.VideoShortEvent
-import com.vitorpamplona.quartz.nip71Video.VideoVerticalEvent
-import com.vitorpamplona.quartz.nip84Highlights.HighlightEvent
-import com.vitorpamplona.quartz.nip88Polls.poll.PollEvent
-import com.vitorpamplona.quartz.nip88Polls.response.PollResponseEvent
-import com.vitorpamplona.quartz.nip99Classifieds.ClassifiedsEvent
-import com.vitorpamplona.quartz.nipA0VoiceMessages.VoiceEvent
-import com.vitorpamplona.quartz.nipA0VoiceMessages.VoiceReplyEvent
-import kotlin.math.min
-
-val HomePostsNewThreadKinds1 =
-    listOf(
-        TextNoteEvent.KIND,
-        RepostEvent.KIND,
-        GenericRepostEvent.KIND,
-        ClassifiedsEvent.KIND,
-        LongTextNoteEvent.KIND,
-        HighlightEvent.KIND,
-        ZapPollEvent.KIND,
-        PollEvent.KIND,
-        WikiNoteEvent.KIND,
-        AttestationEvent.KIND,
-        NipTextEvent.KIND,
-        PictureEvent.KIND,
-        VideoNormalEvent.KIND,
-        VideoShortEvent.KIND,
-        VideoHorizontalEvent.KIND,
-        VideoVerticalEvent.KIND,
-    )
-
-val HomePostsNewThreadKinds2 =
-    listOf(
-        InteractiveStoryPrologueEvent.KIND,
-        ChessGameEvent.KIND,
-        LiveChessGameChallengeEvent.KIND,
-        LiveChessGameEndEvent.KIND,
-        BirdDetectionEvent.KIND,
-        BirdexEvent.KIND,
-        TorrentEvent.KIND,
-    )
-
-val HomePostsConversationKinds =
-    listOf(
-        LiveActivitiesChatMessageEvent.KIND,
-        CommentEvent.KIND,
-        LiveActivitiesEvent.KIND,
-        EphemeralChatEvent.KIND,
-        VoiceEvent.KIND,
-        VoiceReplyEvent.KIND,
-        PollResponseEvent.KIND,
-    )
-
-fun filterNewHomePostsByAuthors(
-    relay: NormalizedRelayUrl,
-    authors: Set<HexKey>,
-    since: Long? = null,
-): List<RelayBasedFilter> {
-    val authorList = authors.sorted()
-
-    return listOf(
-        RelayBasedFilter(
-            relay = relay,
-            filter =
-                ExplainedFilter(
-                    purpose = SubPurpose.HOME_FEED,
-                    kinds = HomePostsNewThreadKinds1,
-                    authors = authorList,
-                    limit = min(authorList.size * 10, 500),
-                    since = since,
-                ),
-        ),
-        RelayBasedFilter(
-            relay = relay,
-            filter =
-                ExplainedFilter(
-                    purpose = SubPurpose.HOME_FEED,
-                    kinds = HomePostsNewThreadKinds2,
-                    authors = authorList,
-                    limit = min(authorList.size * 10, 5),
-                    since = since,
-                ),
-        ),
-    )
-}
-
-fun filterReplyHomePostsByAuthors(
-    relay: NormalizedRelayUrl,
-    authors: Set<HexKey>,
-    since: Long? = null,
-): List<RelayBasedFilter> {
-    val authorList = authors.sorted()
-
-    return listOf(
-        RelayBasedFilter(
-            relay = relay,
-            filter =
-                ExplainedFilter(
-                    purpose = SubPurpose.HOME_FEED,
-                    kinds = HomePostsConversationKinds,
-                    authors = authorList,
-                    limit = min(authorList.size * 10, 500),
-                    since = since,
-                ),
-        ),
-    )
-}
-
-fun filterHomePostsByAuthors(
-    authorSet: AuthorsTopNavPerRelayFilterSet,
-    since: SincePerRelayMap?,
-    sinceBoundaryNew: Long?,
-    sinceBoundaryReply: Long?,
-): List<RelayBasedFilter> {
-    if (authorSet.set.isEmpty()) return emptyList()
-
-    return authorSet.set
-        .mapNotNull {
-            if (it.value.authors.isEmpty()) {
-                null
-            } else {
-                filterNewHomePostsByAuthors(
-                    relay = it.key,
-                    authors = it.value.authors,
-                    since = since?.get(it.key)?.time ?: sinceBoundaryNew,
-                ) +
-                    filterReplyHomePostsByAuthors(
-                        relay = it.key,
-                        authors = it.value.authors,
-                        since = since?.get(it.key)?.time ?: sinceBoundaryReply,
-                    )
-            }
-        }.flatten()
-}
-
-fun filterHomePostsByAuthors(
-    authorSet: MutedAuthorsTopNavPerRelayFilterSet,
-    since: SincePerRelayMap?,
-    sinceBoundaryNew: Long?,
-    sinceBoundaryReply: Long?,
-): List<RelayBasedFilter> {
-    if (authorSet.set.isEmpty()) return emptyList()
-
-    return authorSet.set
-        .mapNotNull {
-            if (it.value.authors.isEmpty()) {
-                null
-            } else {
-                filterNewHomePostsByAuthors(
-                    relay = it.key,
-                    authors = it.value.authors,
-                    since = since?.get(it.key)?.time ?: sinceBoundaryNew,
-                ) +
-                    filterReplyHomePostsByAuthors(
-                        relay = it.key,
-                        authors = it.value.authors,
-                        since = since?.get(it.key)?.time ?: sinceBoundaryReply,
-                    )
-            }
-        }.flatten()
-}
