+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/threadview/datasources/ThreadFilterAssemblerSubscription.kt
@@ -24,6 +24,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import com.vitorpamplona.amethyst.commons.model.Note
 import com.vitorpamplona.amethyst.commons.relayClient.subscriptions.LifecycleAwareKeyDataSourceSubscription
+import com.vitorpamplona.amethyst.commons.relayClient.thread.ThreadFilterAssembler
+import com.vitorpamplona.amethyst.commons.relayClient.thread.ThreadQueryState
 import com.vitorpamplona.amethyst.model.Account
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.AccountViewModel
 import com.vitorpamplona.quartz.nip01Core.core.HexKey
@@ -48,7 +50,7 @@ fun ThreadFilterAssemblerSubscription(
     // even if they are tracking the same tag.
     val state =
         remember(eventId) {
-            ThreadQueryState(eventId, account)
+            ThreadQueryState(eventId, account, account.followPlusAllMineWithSearch.flow)
         }
 
     LifecycleAwareKeyDataSourceSubscription(state, filterAssembler)
