+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/relayClient/home/HomeFilterAssembler.kt
@@ -1,0 +1,54 @@
+/*
+ * Copyright (c) 2025 Vitor Pamplona
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
+ * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.vitorpamplona.amethyst.commons.relayClient.home
+
+import androidx.compose.runtime.Stable
+import com.vitorpamplona.amethyst.commons.model.HomeFeedType
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.topNavFeeds.IFeedTopNavPerRelayFilterSet
+import com.vitorpamplona.amethyst.commons.model.topNavFeeds.TopFilter
+import com.vitorpamplona.amethyst.commons.relayClient.home.nip65Follows.HomeOutboxEventsEoseManager
+import com.vitorpamplona.amethyst.commons.relayClient.topNavFeeds.TopNavFeedFilterAssembler
+import com.vitorpamplona.amethyst.commons.relayClient.topNavFeeds.TopNavFeedQueryState
+import com.vitorpamplona.amethyst.commons.ui.feeds.FeedContentState
+import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.StateFlow
+
+/**
+ * Key for the home feed: the top-nav selection, the new-threads and replies floors that bound
+ * `since`, and the Settings › Home content-type toggles whose disabled groups are stripped from
+ * every home REQ.
+ */
+@Stable
+class HomeQueryState(
+    account: IAccount,
+    listName: StateFlow<TopFilter>,
+    followsPerRelay: StateFlow<IFeedTopNavPerRelayFilterSet>,
+    scope: CoroutineScope,
+    val newThreads: FeedContentState,
+    val replies: FeedContentState,
+    val enabledHomeFeedTypes: StateFlow<Set<HomeFeedType>>,
+) : TopNavFeedQueryState(account, listName, followsPerRelay, scope, listOf(newThreads, replies))
+
+class HomeFilterAssembler(
+    client: INostrClient,
+) : TopNavFeedFilterAssembler<HomeQueryState>({ keys -> listOf(HomeOutboxEventsEoseManager(client, keys)) })
